Teste de Middleware para autenticação e autenticação no controller

+++ app/Http/Controllers/ProjectController.php
@@ -8,6 +8,7 @@ use CodeProject\Services\ProjectService;
 use Illuminate\Http\Request;
 
 use CodeProject\Http\Requests;
+use LucaDegasperi\OAuth2Server\Facades\Authorizer;
 
 class ProjectController extends Controller
 {
@@ -37,7 +38,7 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        return $this->repository->all();
+        return $this->repository->findWhere(['owner_id' => \Authorizer::getResourceOwnerId()]);
     }
 
     /**
@@ -53,7 +54,7 @@ class ProjectController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  Request  $request
+     * @param  Request $request
      * @return Response
      */
     public function store(Request $request)
@@ -64,18 +65,22 @@ class ProjectController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return Response
      */
     public function show($id)
     {
+        if ($this->checkProjectOwner($id)==false){
+            return ['error' => 'Access forbidden'];
+        }
+
         return $this->repository->find($id);
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return Response
      */
     public function edit($id)
@@ -86,29 +91,44 @@ class ProjectController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  Request  $request
-     * @param  int  $id
+     * @param  Request $request
+     * @param  int $id
      * @return Response
      */
     public function update(Request $request, $id)
     {
+        if ($this->checkProjectOwner($id)==false){
+            return ['error' => 'Access forbidden'];
+        }
+
         $this->service->update($request->all(), $id);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return Response
      */
     public function destroy($id)
     {
 
+        if ($this->checkProjectOwner($id)==false){
+            return ['error' => 'Access forbidden'];
+        }
+
         return ['success' => $this->repository->delete($id)];
 
-            //['success' => Client::find($id)->delete()];
+        //['success' => Client::find($id)->delete()];
 
 
+    }
+
+    private function checkProjectOwner($projectId)
+    {
+        $userId = \Authorizer::getResourceOwnerId();
+
+        return $this->repository->isOwner($projectId, $userId);
 
     }
 }
